Create utils for playgrounds
