+++ playground/relay-kit/usdc-transfer-4337-sponsored.ts
@@ -1,5 +1,5 @@
-import { ethers } from 'ethers'
 import { Safe4337Pack } from '@safe-global/relay-kit'
+import { waitForOperationToFinish, generateTransferCallData } from '../utils'
 
 // Safe owner PK
 const PRIVATE_KEY = ''
@@ -90,28 +90,7 @@ async function main() {
     executable: signedSafeOperation
   })
 
-  console.log(`https://jiffyscan.xyz/userOpHash/${userOperationHash}?network=${CHAIN_NAME}`)
-
-  let userOperationReceipt = null
-  while (!userOperationReceipt) {
-    await new Promise((resolve) => setTimeout(resolve, 2000))
-    userOperationReceipt = await safe4337Pack.getUserOperationReceipt(userOperationHash)
-  }
-
-  console.group('User Operation Receipt and hash')
-  console.log('User Operation Receipt', userOperationReceipt)
-  console.log(
-    'User Operation By Hash',
-    await safe4337Pack.getUserOperationByHash(userOperationHash)
-  )
-  console.groupEnd()
+  await waitForOperationToFinish(userOperationHash, CHAIN_NAME, safe4337Pack)
 }
 
 main()
-
-const generateTransferCallData = (to: string, value: bigint) => {
-  const functionAbi = 'function transfer(address _to, uint256 _value) returns (bool)'
-  const iface = new ethers.Interface([functionAbi])
-
-  return iface.encodeFunctionData('transfer', [to, value])
-}
